+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-### 0.2.6
+### 0.3.0
 
 * Fix default formatter not defined error.
 * Add ellipsis to options that don't fit into the screen.
 * Re-render on terminal resize.
+* Scroll if options don't fit into the terminal.
 
 ### 0.2.5